Restructuring EventListener files
