+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/game/Game.java
@@ -28,8 +28,8 @@ import de.butzlabben.missilewars.game.timer.GameTimer;
 import de.butzlabben.missilewars.game.timer.LobbyTimer;
 import de.butzlabben.missilewars.game.timer.Timer;
 import de.butzlabben.missilewars.inventory.OrcItem;
-import de.butzlabben.missilewars.listener.GameBoundListener;
 import de.butzlabben.missilewars.listener.game.EndListener;
+import de.butzlabben.missilewars.listener.game.GameBoundListener;
 import de.butzlabben.missilewars.listener.game.GameListener;
 import de.butzlabben.missilewars.listener.game.LobbyListener;
 import de.butzlabben.missilewars.util.MotdManager;
