+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/listener/game/GameBoundListener.java
@@ -16,7 +16,7 @@
  * along with MissileWars.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.butzlabben.missilewars.listener;
+package de.butzlabben.missilewars.listener.game;
 
 import de.butzlabben.missilewars.MissileWars;
 import de.butzlabben.missilewars.game.Game;
