+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/listener/game/GameListener.java
@@ -21,7 +21,6 @@ package de.butzlabben.missilewars.listener.game;
 import de.butzlabben.missilewars.MessageConfig;
 import de.butzlabben.missilewars.game.Game;
 import de.butzlabben.missilewars.game.GameResult;
-import de.butzlabben.missilewars.listener.GameBoundListener;
 import de.butzlabben.missilewars.util.version.VersionUtil;
 import de.butzlabben.missilewars.wrapper.abstracts.arena.FallProtectionConfiguration;
 import de.butzlabben.missilewars.wrapper.event.PlayerArenaJoinEvent;
