+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/listener/game/LobbyListener.java
@@ -21,7 +21,6 @@ package de.butzlabben.missilewars.listener.game;
 import de.butzlabben.missilewars.MessageConfig;
 import de.butzlabben.missilewars.game.Game;
 import de.butzlabben.missilewars.inventory.VoteInventory;
-import de.butzlabben.missilewars.listener.GameBoundListener;
 import de.butzlabben.missilewars.util.version.VersionUtil;
 import de.butzlabben.missilewars.wrapper.event.PlayerArenaJoinEvent;
 import de.butzlabben.missilewars.wrapper.event.PlayerArenaLeaveEvent;
